+++ CHANGELOG.md
@@ -17,6 +17,8 @@
   - Display format of time durations can be `hh:mm:ss`, `hh:mm`, and `mm:ss`
   - Date variables can now be additionally formatted in dd-MMM-yyyy format, ex: 08-FEB-2016
   - Calculation fields can now be set as hidden during data entry
+- **Notification Changes**
+  - Adverse event changes, sheet comments, and tablet handoff completions now have in-app notifications
 - **Project Changes**
   - Added script to copy projects as templates to allow user testing of subject creation and sheet entry
   - The main project page now displays an improved subjects index